Seperated AnimalList
